+++ converters/pdf/src/pdf_visitor.rs
@@ -3364,10 +3364,12 @@ impl<'a, 'd, 'm> PdfVisitor<'a, 'd, 'm> {
                     Ok(())
                 })
             }
-            XrefDisplay::FullEmphasized(prefix, inlines, _scope) => {
+            XrefDisplay::Emphasized(prefix, inlines, _scope) => {
                 self.write_labelled_link(target, |visitor| {
-                    visitor.write_text_expr(&prefix);
-                    visitor.write_text_expr(", ");
+                    if let Some(prefix) = prefix {
+                        visitor.write_text_expr(&prefix);
+                        visitor.write_text_expr(", ");
+                    }
                     visitor.write_quoted_span(traversal, None, None, "#emph[", inlines, "]")?;
                     Ok(())
                 })